requirements.txt
flask==3.1.2
mysql-connector-python==9.4.0
pydantic==2.11.7
Werkzeug==3.1.3
